+++ src/main/java/dd/blue/model/BlueCoderRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlueCoderRepository extends CrudRepository<BlueCoder, Long> {
+public interface BlueCoderRepository extends CrudRepository<BlueCoder, Integer> {
 
 }
