+++ src/main/java/dd/blue/model/BlueRoleRepository.java
@@ -2,6 +2,6 @@ package dd.blue.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BlueRoleRepository extends CrudRepository<BlueRole, Long>{
+public interface BlueRoleRepository extends CrudRepository<BlueRole, Integer>{
 
 }
