+++ src/main/java/dd/blue/model/BlueTeamRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlueTeamRepository extends CrudRepository<BlueTeam, Long> {
+public interface BlueTeamRepository extends CrudRepository<BlueTeam, Integer> {
 }
